+++ src/nitrate/__init__.py
@@ -13,4 +13,4 @@ This file only exists to make Python's package management tools happy,
 and to hold the version number.
 """
 
-__version__ = "1.2.0"
+__version__ = "1.3.0"