+++ src/nitrate/comments.py
@@ -1,16 +1,80 @@
 """
 Code for dealing with Flickr comments.
 
-In particular, it has some code shared across our
+In particular, some of Flickr's auto-link detection can be a bit wonky,
+e.g. it detects Wikipedia URLs at the wrong point, so we have some code
+that fixes it.  This module allows us to apply similar fixes across
+all of our projects.
+
+Callers only need one function from this file: ``fix_comment_text``.
 """
 
+from collections.abc import Iterator
 import re
 import typing
 from xml.etree import ElementTree as ET
 
 import httpx
+import hyperlink
 
 from .xml import find_required_elem
+
+
+def fix_comment_text(comment_text: str) -> str:
+    """
+    Fix broken links and other markup issues in comment text.
+    """
+    try:
+        comment_text = fix_wikipedia_links(comment_text)
+    except Exception:
+        pass
+
+    return comment_text
+
+
+class Link(typing.TypedDict):
+    """
+    Represents an <a> tag found in a block of text.
+    """
+
+    raw_markup: str
+    href: str
+    display_text: str
+    suffix: str | None
+
+
+def find_links(comment_text: str) -> Iterator[Link]:
+    """
+    Find links in a block of text.
+
+    For this function, we rely on the fact that Flickr auto-links in
+    a very particular way, for example:
+
+        <a href="https://example.com" rel="noreferrer nofollow">example.com</a>
+
+    It might be safer to use a proper parser like BeautifulSoup,
+    but a Flickr comment isn't a complete HTML document and using
+    regex makes it easier to do minimally invasive replacements.
+
+    Note: some older comments omit the ``noreferrer`` attribute.
+
+    """
+    # This regex is designed to match all <a> tags in the form above,
+    # then capture everything after that up to the next bit of whitespace.
+    for m in re.finditer(
+        r'<a href="(?P<href>[^"]+)" '
+        r'rel="(?:noreferrer )?nofollow">'
+        r"(?P<display_text>[^<]+)"
+        r"</a>"
+        r"(?P<suffix>[^\s]*)",
+        comment_text,
+    ):
+        yield {
+            "raw_markup": m.group(0),
+            "href": m.group("href"),
+            "display_text": m.group("display_text"),
+            "suffix": m.group("suffix") or None,
+        }
 
 
 def fix_wikipedia_links(comment_text: str) -> str:
@@ -26,54 +90,77 @@ def fix_wikipedia_links(comment_text: str) -> str:
     It will omit the final period from the link, which means it goes to
     the wrong page.
 
+    This is a particular issue for Flickr Commons, where people leave
+    a lot of comments linking to Wikipedia pages.
+
     This function will fix the Wikipedia links auto-detected by Flickr.
     It moves any trailing punctuation that's part of the link inside
-    the <a>.  We aren't changing the text of the comment, just the
-    auto-detected HTML markup.
+    the <a>.  We should never change the text of the comment, just move
+    bits in/out of the <a>.
 
     See https://github.com/Flickr-Foundation/commons.flickr.org/issues/297
     """
-    for m in re.finditer(
-        r'<a href="https://en\.wikipedia\.org/wiki/(?P<slug1>[A-Za-z_]+)"'
-        r' rel="noreferrer nofollow">'
-        r"en\.wikipedia\.org/wiki/(?P<slug2>[A-Za-z_]+)"
-        r"</a>"
-        # The suffix can be:
-        #
-        #   - a full stop, which stands alone
-        #   - some text in parentheses, which is a disambiguation string.
-        #     This must be some non-empty text in parens.
-        #
-        r"(?P<suffix>\.|\([^\)]+\))",
-        comment_text,
-    ):
-        print(m)
-        # This is a defensive measure, because it was easier than
-        # getting lookback groups working in the regex.
-        if m.group("slug1") != m.group("slug2"):  # pragma: no cover
+    for link in find_links(comment_text):
+        # If this link is immediately followed by whitespace, we can be
+        # confident we've got the full link.
+        if link["suffix"] is None:
             continue
 
-        orig_title = m.group("slug1")
+        # Parse the URL and get the title of the page
+        #
+        # We're looking for a URL of the form
+        #
+        #     https://en.wikipedia.org/wiki/{title}
+        #
+        # Anything else we should skip.
+        url = hyperlink.parse(link["href"])
+
+        is_english_wikipedia = url.host in {
+            "en.wikipedia.org",
+            "en.m.wikipedia.org",
+            "commons.wikimedia.org",
+        }
+
+        # It's easier to write a regex than enumerate all of these.
+        # e.g. de.wikipedia.org, fr.wikipedia.org
+        is_other_wikipedia = re.match(r"^[a-z]{2}\.wikipedia.org$", url.host)
+
+        if not is_english_wikipedia and not is_other_wikipedia:
+            continue
+
+        if len(url.path) < 2 or url.path[0] != "wiki":
+            continue
+
+        orig_title = url.path[1]
 
         # If there's a Wikipedia page with this exact title, then the
         # link works and we can leave it as-is.
-        if _get_wikipedia_page(orig_title) == "found":
+        if _get_wikipedia_page(host=url.host, title=orig_title) == "found":
             continue
 
         # Otherwise, check to see if there's a page with the suffix
         # added -- and if there does, use that as the new link.
-        alt_title = orig_title + m.group("suffix")
+        #
+        # We don't include the fragment when looking up the title of
+        # the page, but we do add it if we're fixing the URL.
+        if "#" in link["suffix"]:
+            suffix, fragment = link["suffix"].split("#")
+            fragment = f"#{fragment}"
+        else:
+            suffix = link["suffix"]
+            fragment = ""
 
-        print(orig_title)
-        print(alt_title)
+        alt_title = orig_title + suffix
 
-        if _get_wikipedia_page(alt_title) == "found":
+        if _get_wikipedia_page(host=url.host, title=alt_title) == "found":
+            new_url = f"{url.host}/wiki/{alt_title}{fragment}"
+
             comment_text = comment_text.replace(
-                m.group(0),
+                link["raw_markup"],
                 (
-                    f'<a href="https://en.wikipedia.org/wiki/{alt_title}" '
+                    f'<a href="{url.scheme}://{new_url}" '
                     'rel="noreferrer nofollow">'
-                    f"en.wikipedia.org/wiki/{alt_title}</a>"
+                    f"{new_url}</a>"
                 ),
             )
 
@@ -83,7 +170,7 @@ def fix_wikipedia_links(comment_text: str) -> str:
 WikipediaPageStatus = typing.Literal["found", "redirected", "not_found"]
 
 
-def _get_wikipedia_page(title: str) -> WikipediaPageStatus:
+def _get_wikipedia_page(host: str, title: str) -> WikipediaPageStatus:
     """
     Look up a page on Wikipedia and see whether it:
 
@@ -92,8 +179,11 @@ def _get_wikipedia_page(title: str) -> WikipediaPageStatus:
     3.  Isn't found
 
     """
+    if host == "en.m.wikipedia.org":
+        host = "en.wikipedia.org"
+
     resp = httpx.get(
-        "https://en.wikipedia.org/w/api.php",
+        f"https://{host}/w/api.php",
         params={
             "action": "query",
             "prop": "revisions",